% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PageRank.R
\name{pagerk}
\alias{pagerk}
\title{Use R to give a function of PageRank.}
\usage{
pagerk(pages)
}
\arguments{
\item{pages}{is a page is the adjacency matrix of the webpage.}
}
\value{
a vector
}
\description{
PageRank is a Google proprietary algorithm used to measure the importance of a particular webpage relative to other webpages in the search engine index.
}
\details{
This is some descriptio of this function.

Considering and not considering the situation of the damping system.
}
