\name{StatComp21094-package}
\alias{StatComp21094-package}
\docType{package}
\title{
A job of statistical calculation.
}
\description{
Several R and Rcpp function are included in this package for illustration.
}
\author{
Daao Yu
Maintainer: Daao Yu <2351651935@qq.com> }
\keyword{ package }
